Add a search page
On this page the user inputs a course number they would like to find a partner in.
When they select a course from the dropdown, the page should switch to
the Feed with a filtered list of people.

TODO: Feed page, Home page where the user can choose to edit their profile
or find a partner